Check for missing jobs in the web worker
This is done to avoid nasty unhandled errors because of race conditions.

If for some reason (e.g. artificial delays introduced to test the UI) a
command on a job results on that job being deleted from the job registry
while there are some other pending commands, this can result on an error
because that job will not be defined when the pending command runs.

+++ ww.js
@@ -120,6 +120,7 @@ function createJobHandler (file) {
 handlers.processJob = processJobHandler;
 function processJobHandler (jobId) {
     const job = Job.jobRegistry.get(jobId);
+    if (typeof job === 'undefined') return;
 
     if (job.file.size > MAX_FILE_SIZE) {
         postReply(replies.fileTooLarge, jobId);
@@ -133,6 +134,7 @@ function processJobHandler (jobId) {
 handlers.deleteJob = deleteJobHandler;
 function deleteJobHandler (jobId) {
     const job = Job.jobRegistry.get(jobId);
+    if (typeof job === 'undefined') return;
 
     job.delete();
